requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-allauth==65.4.1
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.10
importlib_metadata==8.6.1
Markdown==3.7
pillow==11.1.0
psycopg2==2.9.10
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
